(SIMP-7121) Add libkv to simplib::passgen
- Updated `simplib::passgen` to run in legacy mode or in a libkv mode.
  - When in libkv mode, `simplib:passgen` uses libkv for password persistence.
  - libkv mode is enabled by setting `simplib::passgen::libkv` to `true`
    in hieradata.
  - If you enable libkv mode in a system that already has passwords generated
    via the legacy code, currently, **all passwords will be regenerated**.
  - Added `libkv_options` parameter to `simplib::passgen` for use in libkv mode.
  - Added `gen_timeout_seconds` password option.  Previously this was hardcoded
    to 30 seconds.
- Added simp-libkv as a module dependency.

SIMP-7121 #close
