+++ .fixtures.yml
@@ -1,7 +1,9 @@
 ---
 fixtures:
   repositories:
+    libkv: https://github.com/simp/pupmod-simp-libkv
     stdlib: https://github.com/simp/puppetlabs-stdlib
+
     # This needs to be in place for the rspec-puppet Hiera 5 hook to work
     # No idea why, it may be because Puppet sees a custom backend and loads all
     # of the global parts.
