+++ .ruby-version
@@ -1,1 +1,1 @@
-2.4.4
+2.4.5
