+++ metadata.json
@@ -19,6 +19,10 @@
     {
       "name": "puppetlabs/stdlib",
       "version_requirement": ">= 4.13.1 < 7.0.0"
+    },
+    {
+      "name": "simp/libkv",
+      "version_requirement": ">= 0.7.0 < 1.0.0"
     }
   ],
   "operatingsystem_support": [
