+++ spec/fixtures/hieradata/simplib_passgen_libkv.yaml
@@ -1,0 +1,2 @@
+---
+simplib::passgen::libkv: true
